Broke records section into its own app
