+++ un_project/un_app/urls.py
@@ -4,11 +4,6 @@ from .views import *
 urlpatterns = [
     path('', home, name='home'),  # Default home page
     path('financial-home', financial_home, name='financial_home'),
-    path('records-home', records_home, name='records_home'),
-    path('charter', charter, name='charter'),
-    path('resolutions', resolutions, name='resolutions'),
-    path('court-cases', court_cases, name='court_cases'),
-    path('treaties', treaties, name='treaties'),
     path('buildings/', building_list, name='building_list'),  # http://127.0.0.1:8000/buildings/
     path('players/', player_list, name='player_list'),        # http://127.0.0.1:8000/players/
     path('nation/<str:nation_abbreviation>/', nation_balance_sheet, name='nation_balance_sheet'),  # http://127.0.0.1:8000/nation/XYZ/
