+++ un_project/un_project/urls.py
@@ -22,6 +22,7 @@ urlpatterns = [
     path('', include('website.urls')),
     
     path('minecraft/records/', include('un_app.urls')),   # Main app URLs
+    path('minecraft/records/', include('un_records_app.urls')),   # Main app URLs
 
     path('minecraft/records/un_api/', include('un_api.urls')),  # API URLs
     path('minecraft/player/api/', include('players_api.urls'))
