+++ un_project/un_records_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UnRecordsAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'un_records_app'
