+++ un_project/un_records_app/templates/court_cases.html
@@ -1,6 +1,6 @@
-{% extends 'records_templates/records_base.html' %}
+{% extends 'records_base.html' %}
 
-{% block title %}Resolutions{% endblock %}
+{% block title %}Court Cases{% endblock %}
 
 {% load static %}
 {% load custom_filters %}
@@ -11,7 +11,7 @@
 
 {% block content %}
 <div class="container">
-    <h1>United Nations Resolutions</h1>
+    <h1>United Nations Court Cases</h1>
 
 </div>
 {% endblock %}
