+++ un_project/un_records_app/templates/records_home.html
@@ -1,4 +1,4 @@
-{% extends 'records_templates/records_base.html' %}
+{% extends 'records_base.html' %}
 
 {% load static %}  <!-- Load the static tag library -->
