+++ un_project/un_records_app/templates/treaties.html
@@ -1,6 +1,6 @@
-{% extends 'records_templates/records_base.html' %}
+{% extends 'records_base.html' %}
 
-{% block title %}Court Cases{% endblock %}
+{% block title %}Treaties{% endblock %}
 
 {% load static %}
 {% load custom_filters %}
@@ -11,7 +11,7 @@
 
 {% block content %}
 <div class="container">
-    <h1>United Nations Court Cases</h1>
+    <h1>United Nations Treaties</h1>
 
 </div>
 {% endblock %}
